services/pass_vault/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
